requirements.txt
kubernetes
openstacksdk
